Moved some files around so that npm run build will run correctly - there are still a couple of things that are out of wack here, but I didnt want to move other people's components
