+++ components/Navbar.js
@@ -1,8 +1,8 @@
 import Link from 'next/link';
 import LoginButton from '../src/pages/Login/Login.js';
-import Cookies from '../src/pages/Login/setCookie.js';
+import Cookies from '../helper_functions/setCookie.js';
 import Register from '../src/pages/Login/register.js';
-import Logout from '../src/pages/Login/logOut.js';
+import Logout from '../helper_functions/logOut.js';
 import react, {useEffect, useState} from 'react';
 
 import localFont from 'next/font/local';
