+++ helper_functions/facebookLogin.js
