+++ helper_functions/logOut.js
