+++ helper_functions/loginUser.js
