+++ helper_functions/registrationValidation.js
