+++ helper_functions/resetPassword.js
