+++ helper_functions/setCookie.js
