+++ src/pages/Login/Login.js
@@ -2,10 +2,10 @@ import React, {useState, useEffect} from 'react';
 import { Button, Modal} from 'react-bootstrap';
 import login from '../api/login.js';
 import firebase from '../../../firebase/clientApp.js';
-import facebookLogin from './facebookLogin.js';
-import googleLogin from './googleLogin.js';
-import CreateUser from './newUser.js';
-import LoginUser from './loginUser.js';
+import facebookLogin from '../../../helper_functions/facebookLogin.js';
+import googleLogin from '../../../helper_functions/googleLogin.js';
+import CreateUser from '../../../helper_functions/newUser.js';
+import LoginUser from '../../../helper_functions/loginUser.js';
 import Link from 'next/link';
 
 export default function Login(props1) {
