+++ src/pages/authenticated.js
@@ -1,7 +1,7 @@
 import MyActivities from '../../components/MyActivities.js';
 import CompletedActivites from '../../components/completedActivites.js';
 import Favorites from '../../components/Favorites.js';
-import Cookies from './Login/setCookie.js';
+import Cookies from '../../helper_functions/setCookie.js';
 import {useRouter} from 'next/router';
 import react, { useEffect, useState } from 'react';
 import Image from 'next/image';
